requirements.txt
fastapi
uvicorn
asyncpg
python-dotenv
pydantic
groq
streamlit
requests